+++ tala/nl/rasa/generator.py
@@ -48,11 +48,11 @@ class RasaGenerator(AbstractGenerator):
 
     @property
     def _sortal_entity_template(self):
-        return Template("[{{ grammar_entry }}](sort.{{ value }})")
+        return Template("[{{ grammar_entry }}]({{ ddd }}.sort.{{ value }})")
 
     @property
     def _propositional_entity_template(self):
-        return Template("[{{ grammar_entry }}](predicate.{{ value }})")
+        return Template("[{{ grammar_entry }}]({{ ddd }}.predicate.{{ value }})")
 
     def _generate_examples(self):
         examples = super(RasaGenerator, self)._generate_examples()
@@ -121,16 +121,16 @@ class RasaGenerator(AbstractGenerator):
             "synonyms": synonyms,
         }
 
-    def _create_intent_samples(self, grammar, intent):
+    def _create_intent_samples(self, grammar, ddd, intent):
         head = intent.text_chunks[0]
         texts = intent.text_chunks[1:]
         try:
-            samples = self._examples_with_individuals(grammar, texts, intent.required_entities, [head])
+            samples = self._examples_with_individuals(grammar, ddd, texts, intent.required_entities, [head])
             return samples
         except UnexpectedPropositionalEntityEncounteredException:
             return []
 
-    def _examples_with_individuals(self, grammar, text_chunks, required_entities, examples_so_far):
+    def _examples_with_individuals(self, grammar, ddd, text_chunks, required_entities, examples_so_far):
         if not text_chunks and not required_entities:
             return examples_so_far
         tail = text_chunks[0]
@@ -138,13 +138,13 @@ class RasaGenerator(AbstractGenerator):
         all_new_examples = []
         for example in examples_so_far:
             if required_entity.is_sortal:
-                new_examples = list(self._examples_from_sortal_individual(grammar, required_entity, example, tail))
+                new_examples = list(self._examples_from_sortal_individual(grammar, ddd, required_entity, example, tail))
                 all_new_examples.extend(new_examples)
             elif required_entity.is_propositional:
                 predicate = self._ddd.ontology.get_predicate(required_entity.name)
                 if predicate.getSort().is_string_sort():
                     new_examples = list(
-                        self._examples_from_propositional_individual(grammar, required_entity, example, tail)
+                        self._examples_from_propositional_individual(grammar, ddd, required_entity, example, tail)
                     )
                     all_new_examples.extend(new_examples)
 
@@ -160,13 +160,13 @@ class RasaGenerator(AbstractGenerator):
                     "Expected either a sortal or propositional required entity but got a %s" %
                     required_entity.__class__.__name__
                 )
-        return self._examples_with_individuals(grammar, text_chunks[1:], required_entities[1:], all_new_examples)
+        return self._examples_with_individuals(grammar, ddd, text_chunks[1:], required_entities[1:], all_new_examples)
 
-    def _examples_from_sortal_individual(self, grammar, required_sortal_entity, example_so_far, tail):
+    def _examples_from_sortal_individual(self, grammar, ddd, required_sortal_entity, example_so_far, tail):
         sort = self._ddd.ontology.get_sort(required_sortal_entity.name)
         individuals = self._individual_grammar_entries_samples(grammar, sort)
         template = self._entity_template_of(sort)
-        return self._examples_from_individuals(template, sort.get_name(), individuals, example_so_far, tail)
+        return self._examples_from_individuals(template, ddd, sort.get_name(), individuals, example_so_far, tail)
 
     def _individual_grammar_entries_samples(self, grammar, sort):
         if sort.is_builtin():
@@ -177,7 +177,7 @@ class RasaGenerator(AbstractGenerator):
         examples = self._language_examples.get_builtin_sort_examples(sort)
         return [[entry] for entry in examples]
 
-    def _examples_from_propositional_individual(self, grammar, required_propositional_entity, example_so_far, tail):
+    def _examples_from_propositional_individual(self, grammar, ddd, required_propositional_entity, example_so_far, tail):
         predicate_name = required_propositional_entity.name
         predicate = self._ddd.ontology.get_predicate(predicate_name)
         sort = predicate.getSort()
@@ -185,16 +185,16 @@ class RasaGenerator(AbstractGenerator):
         predicate_specific_samples = self._string_examples_of_predicate(grammar, predicate)
         individuals.extend([[predicate_specific_sample] for predicate_specific_sample in predicate_specific_samples])
         return self._examples_from_individuals(
-            self._propositional_entity_template, predicate_name, individuals, example_so_far, tail
+            self._propositional_entity_template, ddd, predicate_name, individuals, example_so_far, tail
         )
 
     def _string_examples_of_predicate(self, grammar, predicate):
         return grammar.strings_of_predicate(predicate.get_name())
 
-    def _examples_from_individuals(self, template, identifier, individuals_grammar_entries, example_so_far, tail):
+    def _examples_from_individuals(self, template, ddd, identifier, individuals_grammar_entries, example_so_far, tail):
         for grammar_entries in individuals_grammar_entries:
             for grammar_entry in grammar_entries:
-                entity = template.render(grammar_entry=grammar_entry, value=identifier)
+                entity = template.render(grammar_entry=grammar_entry, ddd=ddd, value=identifier)
                 example = self._extend_example(entity, example_so_far, tail)
                 yield example
 
@@ -204,10 +204,10 @@ class RasaGenerator(AbstractGenerator):
         tail = tail or ""
         return "".join([head, entity, tail])
 
-    def _create_sortal_answer_samples(self, grammar, sort, intent_templates):
+    def _create_sortal_answer_samples(self, grammar, ddd, sort, intent_templates):
         for grammar_entries in self._individual_grammar_entries_samples(grammar, sort):
             examples = self._examples_of_individual(
-                grammar_entries, sort.get_name(), intent_templates, self._entity_template_of(sort)
+                grammar_entries, ddd, sort.get_name(), intent_templates, self._entity_template_of(sort)
             )
             for example in examples:
                 yield example
@@ -218,13 +218,13 @@ class RasaGenerator(AbstractGenerator):
             return self._sortal_entity_template
         return self._builtin_entity_template
 
-    def _examples_of_individual(self, grammar_entries, identifier, intent_templates, entity_template):
+    def _examples_of_individual(self, grammar_entries, ddd, identifier, intent_templates, entity_template):
         for grammar_entry in grammar_entries:
-            examples = self._examples_from_templates(grammar_entry, identifier, intent_templates, entity_template)
+            examples = self._examples_from_templates(grammar_entry, ddd, identifier, intent_templates, entity_template)
             for example in examples:
                 yield example
 
-    def _examples_from_templates(self, grammar_entry, identifier, intent_templates, entity_template):
+    def _examples_from_templates(self, grammar_entry, ddd, identifier, intent_templates, entity_template):
         for intent_template in intent_templates:
-            entity = entity_template.render(grammar_entry=grammar_entry, value=identifier)
+            entity = entity_template.render(grammar_entry=grammar_entry, ddd=ddd, value=identifier)
             yield intent_template.render(name=entity)